Tienda_gatos/requirements.txt
asgiref==3.5.0
autopep8==1.6.0
pillow
pycodestyle==2.8.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
django-crispy-forms
django-rest-framework